+++ game/src/Voider/src/com/spiddekauga/voider/scene/WorldScene.java
@@ -153,7 +153,7 @@ public abstract class WorldScene extends Scene {
 	protected void fixCamera() {
 		float width = Gdx.graphics.getWidth() * Config.Graphics.WORLD_SCALE;
 		// Decrease scale of width depending on height scaled
-		float heightScale = Config.Graphics.HEIGHT_DEFAULT / Gdx.graphics.getHeight();
+		float heightScale = ((float) Config.Graphics.HEIGHT_DEFAULT) / Gdx.graphics.getHeight();
 		width *= heightScale;
 		float height = Config.Graphics.HEIGHT_DEFAULT * Config.Graphics.WORLD_SCALE;
